+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceConfigurationPolicyWindows10/MSFT_IntuneDeviceConfigurationPolicyWindows10.psm1
@@ -1980,6 +1980,13 @@ function Get-TargetResource
         $assignmentResult = @()
         foreach ($assignmentEntry in $AssignmentsValues)
         {
+            $Group = Get-MgGroup -GroupId $assignmentEntry.Target.AdditionalProperties.groupId -ErrorAction SilentlyContinue
+            $GroupDisplayName = $null
+            if ($Group.Count -eq 1)
+            {
+                $GroupDisplayName = $Group.DisplayName
+            }
+
             $assignmentValue = @{
                 dataType                                   = $assignmentEntry.Target.AdditionalProperties.'@odata.type'
                 deviceAndAppManagementAssignmentFilterType = $(if ($null -ne $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType)
@@ -1988,6 +1995,7 @@ function Get-TargetResource
                     })
                 deviceAndAppManagementAssignmentFilterId   = $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterId
                 groupId                                    = $assignmentEntry.Target.AdditionalProperties.groupId
+                groupDisplayName                           = $GroupDisplayName
             }
             $assignmentResult += $assignmentValue
         }
@@ -3323,9 +3331,68 @@ function Set-TargetResource
             -DeviceConfigurationId $currentInstance.Id `
             -BodyParameter $UpdateParameters
         $assignmentsHash = @()
-        foreach ($assignment in $Assignments)
+        foreach ($assignment in $assignments)
         {
-            $assignmentsHash += Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+            if ($Assignment.dataType -eq "#microsoft.graph.groupAssignmentTarget" -or `
+                $Assignment.dataType -eq "#microsoft.graph.exclusionGroupAssignmentTarget")
+            {
+                if (![string]::IsNullOrEmpty($Assignment.groupId))
+                {
+                    $Group = Get-MgGroup -GroupId $Assignment.groupId -ErrorAction SilentlyContinue
+                    $GroupId = $Assignment.groupId
+                }
+                else
+                {
+                    $Group = $null
+                    $GroupId = "null"
+                }
+
+                if ($Group.Count -eq 0)
+                {
+                    $Message = "Could not find assignment group with id {0}, trying with display name" -f $GroupId
+                    Write-Verbose -Message $Message
+
+                    if (![string]::IsNullOrEmpty($Assignment.groupDisplayName))
+                    {
+                        $Message = "Checking for the assignment group '{0}'" -f $Assignment.groupDisplayName
+                        Write-Verbose -Message $Message
+
+                        $Filter = "displayName eq '{0}'" -f $Assignment.groupDisplayName
+                        $Group = Get-MgGroup -Filter $Filter -ErrorAction SilentlyContinue
+                        if ($Group.Count -eq 1)
+                        {
+                            $Message = "Found assignment group '{0}' with id '{1}'" -f $Group.DisplayName, $Group.Id
+                            Write-Verbose -Message $Message
+
+                            $Assignment.groupId = $Group.Id
+                        }
+                        else
+                        {
+                            if ([string]::IsNullOrEmpty($Assignment.groupId))
+                            {
+                                $Message = "Could not find assignment group, skipping"
+                                continue
+                            }
+
+                            $Message = "Could not find assignment group '{0}', instead use group with id '{1}'" -f $Assignment.groupDisplayName, $Assignment.groupId
+                            Write-Verbose -Message $Message
+                        }
+                    }
+                    else
+                    {
+                        $Message = "Could not find assignment group, skipping"
+                        Write-Verbose -Message $Message
+                        continue
+                    }
+                }
+            }
+
+            $assignmentHash = Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+            if (![string]::IsNullOrEmpty($Assignment.groupDisplayName))
+            {
+                $assignmentHash.Remove("groupDisplayName") | Out-Null
+            }
+            $assignmentsHash += $assignmentHash
         }
         Update-DeviceConfigurationPolicyAssignment `
             -DeviceConfigurationPolicyId $currentInstance.id `
@@ -4810,7 +4877,7 @@ function Export-TargetResource
             }
             if ($Results.Assignments)
             {
-                $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString -ComplexObject $Results.Assignments -CIMInstanceName DeviceManagementConfigurationPolicyAssignments
+                $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString -ComplexObject $Results.Assignments -CIMInstanceName IntuneDeviceConfigurationPolicyWindows10Assignments
                 if ($complexTypeStringResult)
                 {
                     $Results.Assignments = $complexTypeStringResult